Move authn context to istio/authn for sharing between filters.
